Re-import Latin glyphs
Previously, the glyphs to replace were actually replaced.
Then the rest of the needed glyphs was imported. This led
to there being, for example, two A glyphs (with different
names) since the import was done in two passes, the second
pass did not know to reuse the first A glyph.

So this time, all the Latin glyphs (except for those needing
to be kept) were removed, and then (in one pass) the remaining
Latin glyphs were imported.